Added the to_retrain method to the Dwellers' class.
The method returns a detailed list of Dwellers whose current max
health is significantly lower than it should be for that level.
The method allows the client to pass a percentage of the max health
that a Dweller must have with respect to their potential to be
excluded from the list.